Crate nft

Non-Fungible Token implementation with JSON serialization. NOTES:

The maximum balance value is limited by U128 (2**128 - 1).

JSON calls should pass U128 as a base-10 string. E.g. “100”.

The contract optimizes the inner trie structure by hashing account IDs. It will prevent some abuse of deep tries. Shouldn’t be an issue, once NEAR clients implement full hashing of keys.

The contract tracks the change in storage before and after the call. If the storage increases, the contract requires the caller of the contract to attach enough deposit to the function call to cover the storage cost. This is done to prevent a denial of service attack on the contract by taking all available storage. If the storage decreases, the contract will issue a refund for the cost of the released storage. The unused tokens from the attached deposit are also refunded, so it’s safe to attach more deposit than required.

To prevent the deployed contract from being modified or deleted, it should not have any access keys on its account.

Structs

BaseMetadata

This struct needs to be passed in mint method

Contract

Functions

new

new_default_meta

Initializes the contract owned by owner_id with default metadata (for example purposes only).

nft_approve

nft_is_approved

nft_metadata

nft_mint

Mint a new token with ID=token_id belonging to token_owner_id.

nft_resolve_transfer

nft_revoke

nft_revoke_all

nft_supply_for_owner

nft_token

nft_tokens

nft_tokens_for_owner

nft_total_supply

nft_transfer

nft_transfer_call