Struct nft::Contract

pub struct Contract { /* fields omitted */ }

Implementations

impl Contract

pub fn new_default_meta(owner_id: AccountId) -> Self

Initializes the contract owned by owner_id with default metadata (for example purposes only).

pub fn new(owner_id: AccountId, metadata: NFTContractMetadata) -> Self

pub fn nft_mint(
&mut self,
token_owner_id: AccountId,
token_metadata: BaseMetadata
) -> Token

Mint a new token with ID=token_id belonging to token_owner_id.

Since this example implements metadata, it also requires per-token metadata to be provided in this call. self.tokens.mint will also require it to be Some, since StorageKey::TokenMetadata was provided at initialization.

self.tokens.mint will enforce predecessor_account_id to equal the owner_id given in initialization call to new.